Decoupled debugging tools from main.py. Expanded debugging tool functionality and allow running SPMP_debug.py to assist in test-case generation without flipping bools in main.py. README.md established. Minor typos and preparing for addition of Dynamic Pattern feature.
