+++ README.md
@@ -1,2 +1,63 @@
-# StockMaximization
-Stock Maximization Problem using Exhaustive Search Algorithm
+<div align="center">
+
+# Stock Maximization
+
+Stock Maximization Problem (also known as Knapsack Problem) solved through Exhaustive Algorithm and Dynamic Pattern.
+
+</div>
+
+## Installation
+```sh
+pip install more-itertools
+```
+(or)
+```sh
+pip3 install more-itertools
+```
+
+## Disclaimer
+There may be instances where the program may not work correctly. To avoid this, you may create a local virtual environment and pip install within the env. Activate the env when running the program.
+
+```sh
+python3 -m venv envStockMaximization
+```
+
+## Usage Example
+Run main.py on the terminal or through an IDE after installing the necessary module(s).
+
+Additionally, a built-in debugging assistance tool may be run. In the debugger tool file, toggle-able booleans may be used to define scope. To adjust input and output file names, refer to the settings file.
+
+User may generate a new input file, overwrite any existing input file, quantity of test cases, and the min-max values per each parameter. Specifications for each parameter are listed within the debugger tool file. These may assist in creating test cases.
+```sh
+python3 SPMP_debug.py
+```
+
+### Sample Input
+Trailing white spaces and empty lines within an input file are permitted. Valid entry of zero available stocks should be written as "[]." The formatting is as follows:
+> 8
+>
+> [[10, 11], [6, 8], [9, 10], [6, 6], [6, 3], [3, 12], [3, 10], [9, 3]]
+>
+> 12
+>
+> 0
+>
+> []
+>
+> 5
+>
+### Sample Output
+> 30
+>
+> 0
+>
+
+## Metadata
+P. Tang - pattontanges@csu.fullerton.edu
+[https://github.com/PTanges/StockMaximization/]
+
+Submitted to Professor Dsouza as per California State University, Fullerton's Bachelor of Science, Computer Science undergraduate program.
+
+Course: CPSC 335. Section 02.
+
+Deadline of submission: 12/03/2023 via Canvas in a zipped folder.
