+++ SPMP_SETTINGS.py
@@ -1,10 +1,4 @@
 # Constants
 inputFileName = "stocks.txt"  # include file-name extension
-outputFileName = "output.txt"
-
-# Allow generation of a new stocks.txt file if the file isn't located
-allowFailsafe = True
-
-# Allows the program to write over either an existing inputFileName or a non-existing file
-# Allows dev to continually generate the file without manual deletion
-allowFileOverwrite = False
+outputFileNameEP = "output_EP.txt"
+outputFileNameDP = "output_DP.txt"
